+++ src/api/Ledger.js
@@ -4,13 +4,14 @@ import type { Currency } from '@ledgerhq/live-common/lib/types'
 const BASE_URL = process.env.LEDGER_REST_API_BASE || 'https://api.ledgerwallet.com/'
 
 const mapping = {
-  bch: 'abc',
-  etc: 'ethc',
+  bitcoin_cash: 'abc',
+  ethereum_classic: 'ethc',
+  ethereum_testnet: 'eth_testnet',
 }
 
 export const currencyToFeeTicker = (currency: Currency) => {
   const tickerLowerCase = currency.ticker.toLowerCase()
-  return mapping[tickerLowerCase] || tickerLowerCase
+  return mapping[currency.id] || tickerLowerCase
 }
 
 export const blockchainBaseURL = (currency: Currency) =>
