+++ src/helpers/getAddressForCurrency/btc.js
