+++ src/helpers/getAddressForCurrency/ethereum.js
