+++ src/helpers/signTransactionForCurrency/ethereum.js
