+++ src/helpers/signTransactionForCurrency/index.js
